# schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PECFF - Passive Email Cryptographic Forensics Framework",
    "description": "Passive network forensics and cryptographic analysis API for SMTP/IMAP/POP3 traffic. Implements pure deterministic NIST SP 800-57 risk scoring, JA3/JA4 TLS fingerprinting, unsupervised ML anomaly detection, and hostile-input hardened parsing.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analyses/{analysis_id}": {
      "get": {
        "tags": [
          "Analyses & Sessions"
        ],
        "summary": "Stream full forensic analysis document incrementally without large RAM buffers",
        "description": "Stream full JSON document chunk by chunk using incremental generator.",
        "operationId": "get_analysis_document_api_v1_analyses__analysis_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/sessions": {
      "get": {
        "tags": [
          "Analyses & Sessions"
        ],
        "summary": "Server-side cursor-based session filtering and sorting (sub-200ms p95 on 100k flows)",
        "description": "Cursor-paginated session search with indexed filtering.",
        "operationId": "list_analysis_sessions_api_v1_analyses__analysis_id__sessions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque pagination cursor token",
              "title": "Cursor"
            },
            "description": "Opaque pagination cursor token"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Page item limit",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page item limit"
          },
          {
            "name": "min_risk",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100.0,
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum risk score filter",
              "title": "Min Risk"
            },
            "description": "Minimum risk score filter"
          },
          {
            "name": "band",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Risk band filter: SECURE, ACCEPTABLE, WEAK, HIGH, CRITICAL",
              "title": "Band"
            },
            "description": "Risk band filter: SECURE, ACCEPTABLE, WEAK, HIGH, CRITICAL"
          },
          {
            "name": "protocol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Mail protocol: SMTP, IMAP, POP3",
              "title": "Protocol"
            },
            "description": "Mail protocol: SMTP, IMAP, POP3"
          },
          {
            "name": "starttls_state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "STARTTLS state: S0_TCP_EST to S4_TLS_READY",
              "title": "Starttls State"
            },
            "description": "STARTTLS state: S0_TCP_EST to S4_TLS_READY"
          },
          {
            "name": "is_anomaly",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ML anomaly flag filter",
              "title": "Is Anomaly"
            },
            "description": "ML anomaly flag filter"
          },
          {
            "name": "ja3",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Client JA3 hash filter",
              "title": "Ja3"
            },
            "description": "Client JA3 hash filter"
          },
          {
            "name": "sni",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Target SNI hostname substring",
              "title": "Sni"
            },
            "description": "Target SNI hostname substring"
          },
          {
            "name": "dst_ip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Destination server IP address",
              "title": "Dst Ip"
            },
            "description": "Destination server IP address"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(risk|anomaly|ts)$",
              "description": "Sort field: risk, anomaly, ts",
              "default": "risk",
              "title": "Sort By"
            },
            "description": "Sort field: risk, anomaly, ts"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "description": "Sort order: asc or desc",
              "default": "desc",
              "title": "Sort Order"
            },
            "description": "Sort order: asc or desc"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPaginatedSessions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CursorPaginatedSessions": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SessionSummarySchema"
            },
            "type": "array",
            "title": "Items",
            "description": "Page session records."
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Opaque cursor string for fetching the next page."
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Whether additional records exist beyond this page."
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "Total matched sessions across all pages."
          }
        },
        "type": "object",
        "required": [
          "items",
          "has_more",
          "total_count"
        ],
        "title": "CursorPaginatedSessions",
        "description": "Cursor-based server-side pagination response."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionSummarySchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique session identifier."
          },
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id",
            "description": "Parent analysis execution ID."
          },
          "client_ip": {
            "type": "string",
            "title": "Client Ip",
            "description": "Client IP address."
          },
          "client_port": {
            "type": "integer",
            "title": "Client Port",
            "description": "Client TCP port."
          },
          "server_ip": {
            "type": "string",
            "title": "Server Ip",
            "description": "Server IP address."
          },
          "server_port": {
            "type": "integer",
            "title": "Server Port",
            "description": "Server TCP port."
          },
          "protocol": {
            "type": "string",
            "title": "Protocol",
            "description": "SMTP, IMAP, or POP3."
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "EXPLICIT or IMPLICIT."
          },
          "starttls_state": {
            "type": "string",
            "title": "Starttls State",
            "description": "Final STARTTLS state (S0-S4)."
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score",
            "description": "NIST SP 800-57 risk score (0-100)."
          },
          "risk_band": {
            "type": "string",
            "title": "Risk Band",
            "description": "SECURE, ACCEPTABLE, WEAK, HIGH, CRITICAL."
          },
          "ja3": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ja3",
            "description": "Client JA3 fingerprint hash."
          },
          "ja4": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ja4",
            "description": "Client JA4 fingerprint."
          },
          "sni": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sni",
            "description": "SNI server hostname."
          },
          "first_seen": {
            "type": "number",
            "title": "First Seen",
            "description": "First packet capture timestamp (epoch)."
          },
          "duration_sec": {
            "type": "number",
            "title": "Duration Sec",
            "description": "TCP session duration in seconds."
          },
          "is_anomaly": {
            "type": "boolean",
            "title": "Is Anomaly",
            "description": "ML anomaly flag.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "analysis_id",
          "client_ip",
          "client_port",
          "server_ip",
          "server_port",
          "protocol",
          "mode",
          "starttls_state",
          "risk_score",
          "risk_band",
          "first_seen",
          "duration_sec"
        ],
        "title": "SessionSummarySchema",
        "description": "Summary record for session listing and cursor pagination."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}